Check for suicide and captures
Basic method for checking if pieces can be captured has been implemented
This includes two helper functions for flipping flags (ie. visited flag)
and for flipping pieces when given a list of pieces. This is used for
capturing pieces. BFS is used for checking if pieces can be captured.
Checking if self can be captured is checking for suicide. TODO: ko rule